+++ lib/wechat_pay/utils/xml_parser.ex
@@ -5,26 +5,38 @@ defmodule WechatPay.Utils.XMLParser do
 
   import SweetXml
 
+  alias WechatPay.Error
+
   @doc """
   Convert a XML string to map
   """
   @spec parse(String.t) :: map
   def parse(data) when is_binary(data) do
-    data
-    |> xpath(~x"/xml/child::*"l)
-    |> Enum.map(fn(node) ->
-      key =
-        node
-        |> xpath(~x"name(.)")
-        |> List.to_atom
+    try do
+      parsed_data =
+        data
+        |> xpath(~x"/xml/child::*"l)
+        |> Enum.map(&parse_node/1)
+        |> Enum.into(%{})
 
-      value =
-        node
-        |> xpath(~x"./text()")
-        |> String.Chars.to_string
+      {:ok, parsed_data}
+    catch
+      :exit, _ ->
+        {:error, %Error{reason: "Malformed XML", type: :malformed_xml}}
+    end
+  end
 
-      {key, value}
-    end)
-    |> Enum.into(%{})
+  defp parse_node(node) do
+    key =
+      node
+      |> xpath(~x"name(.)")
+      |> List.to_atom
+
+    value =
+      node
+      |> xpath(~x"./text()")
+      |> String.Chars.to_string
+
+    {key, value}
   end
 end