+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # master
 
 * Fixes warnings on Elixir 1.4.
-* Added `WechatPay.API.Client.get_sandbox_signkey/0` to get sandbox signkey.
+* Added `WechatPay.API.get_sandbox_signkey/0` to get sandbox signkey.
 * Rename `WechatPay.Plug.Notify` -> `WechatPay.Plug.Callback`, and rewrite the
   flow. Now it's easier to handle callbacks from Wehcat's Payment Gateway.
 * Drop support for loading config from `{:system, ENV}`. It seems not a good